Fixed bug by having the network compare the value of 'previousBlockHash' on the potential block. against the hashed value of the most recent block in the consencus chain. This prevents the double adding. Go figure, it works

+++ bitcoin/miner.js
@@ -156,7 +156,7 @@ class Miner {
         // kill current work
         // validate the block and add it to the chain
         // then start working again
-        this.workerProcess.kill();
+        if (this.workerProcess) this.workerProcess.kill();
         this.blockchain.push(newBlock);
         this.cleanTxPool(newBlock.transactions);
       }
